Move bottom sheet logic to provider
